rename all wp to wordpress and build config tab
