+++ playground/src/views/Editor.vue
@@ -144,7 +144,7 @@ export default class Editor extends Vue {
     ),*/
     new Weblet("Node Pilot", "nodepilot", "nodepilot", "deployment", "vm.png"),
     // new Weblet("QSFS Virtual Machine", "qvm", "qvm", "deployment", "vm.png"),
-    new Weblet("Wordepress", "wp", "wp", "deployment", "wp.png"),
+    new Weblet("Wordepress", "wordpress", "wordpress", "deployment", "wordpress.png"),
 
     new Weblet("Contracts", "contractslist", "", "my account", ""),
     new Weblet("Deployments", "deployedlist", "", "my account", ""),
