+++ src/elements/wordpress/index.ts
@@ -1,1 +1,1 @@
-import "./wp.wc.svelte";
+import "./wordpress.wc.svelte";
